requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9

pytest==8.2.1
pytest-cov==5.0.0
flake8==7.1.0
sqlalchemy
